services/hsm/auth/requirements.txt
sanic==21.9.1
wheel
pycrypto